Clean up KryoAtomicCoder
