+++ scio-core/src/main/scala/com/spotify/scio/coders/serializers/CoderSerializer.scala
@@ -15,7 +15,7 @@
  * under the License.
  */
 
-package com.spotify.scio.coders
+package com.spotify.scio.coders.serializers
 
 import com.esotericsoftware.kryo.Kryo
 import com.esotericsoftware.kryo.io.{Input, Output}
@@ -23,7 +23,7 @@ import com.twitter.chill.KSerializer
 import org.apache.beam.sdk.coders.Coder
 import org.apache.beam.sdk.util.CoderUtils
 
-private class CoderSerializer[T](private val coder: Coder[T]) extends KSerializer[T] {
+private[coders] class CoderSerializer[T](private val coder: Coder[T]) extends KSerializer[T] {
 
   override def write(kser: Kryo, out: Output, obj: T): Unit = {
     val bytes = CoderUtils.encodeToByteArray(coder, obj)
