+++ scio-core/src/main/scala/com/spotify/scio/coders/serializers/KVSerializer.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 Spotify AB.
+ * Copyright 2016 Spotify AB.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,21 +15,24 @@
  * under the License.
  */
 
-package com.spotify.scio.coders
+package com.spotify.scio.coders.serializers
 
 import com.esotericsoftware.kryo.Kryo
 import com.esotericsoftware.kryo.io.{Input, Output}
 import com.twitter.chill.KSerializer
+import org.apache.beam.sdk.values.KV
 
-import java.math.{BigDecimal => JBigDecimal}
+private[coders] class KVSerializer[K, V] extends KSerializer[KV[K, V]] {
 
-private class BigDecimalSerializer extends KSerializer[BigDecimal] {
-  override def read(kryo: Kryo, input: Input, cls: Class[BigDecimal]): BigDecimal = {
-    val jBigDec = kryo.readClassAndObject(input).asInstanceOf[JBigDecimal]
-    BigDecimal(jBigDec)
+  override def write(kser: Kryo, out: Output, obj: KV[K, V]): Unit = {
+    kser.writeClassAndObject(out, obj.getKey)
+    kser.writeClassAndObject(out, obj.getValue)
   }
 
-  override def write(kryo: Kryo, output: Output, obj: BigDecimal): Unit = {
-    kryo.writeClassAndObject(output, obj.bigDecimal)
+  override def read(kser: Kryo, in: Input, cls: Class[KV[K, V]]): KV[K, V] = {
+    val k = kser.readClassAndObject(in).asInstanceOf[K]
+    val v = kser.readClassAndObject(in).asInstanceOf[V]
+    KV.of(k, v)
   }
+
 }
